refactor:자동차이름 유효성체크 책임을 racingCars에 넘긴다.
